Community extension: Original Creator
This extension provides a way to track the origin of a badge when one
organisation creates it for others.

For example, presume we have organisations X, Y and Z. A badge is
created by X and they grant Y and Z the right to issue it.  IssuerClass
of X is stored as the OriginalCreator in the BadgeClass and either Y or
Z becomes the issuer.

Signed-off-by: Antti Koskinen <antti.koskinen@discendum.com>