Add Result.status Provisional
The gap between "OnHold" with its somewhat negative connotation and "Completed" is big enough to squeeze in another possibility, "Provisional". This carries the expectation of an eventual completed result, but perhaps has not passed all administrative checks that will complete the status.

I have some use cases for this for self-claimed badges that are not yet fully confirmed by their issuer, but may be useful to circulate in any case. Consumers SHOULD view this as an incomplete status, but they MAY consider it satisfactory.